+++ src/modules/location-huds/locations/cerulean-skyport/index.js
@@ -2,6 +2,7 @@ import { addHudStyles } from '@utils';
 
 import { addAirshipRandomizer } from '../../shared/airship-randomizer';
 import { addHullClick } from './ledger';
+import { initRaidFavorites } from './raid-favorites';
 
 import fullWidthAirshipStyles from '../floating-islands/full-width-airship.css';
 import styles from './styles.css';
@@ -13,4 +14,5 @@ export default async () => {
   addHudStyles([styles, fullWidthAirshipStyles], 'cerulean-skyport');
   addAirshipRandomizer();
   addHullClick();
+  initRaidFavorites();
 };